Openeye-ification version
