+++ chemprop/data/__init__.py
@@ -1,3 +1,4 @@
 from .data import MoleculeDatapoint, MoleculeDataset
-from .scaffold import scaffold_to_smiles
+# Not supporting for OpenEye
+#from .scaffold import scaffold_to_smiles
 from .scaler import StandardScaler
