+++ chemprop/features/OEMolGraph.py
@@ -1,0 +1,166 @@
+from openeye.oechem import *
+
+# Stuff for typing:
+from argparse import Namespace
+from typing import List, Tuple, Union
+
+MAX_ATOMIC_NUM = 100
+ATOM_FEATURES = {
+    'atomic_num': list(range(MAX_ATOMIC_NUM)),
+    'degree': [0, 1, 2, 3, 4, 5],
+    'formal_charge': [-1, -2, 1, 2, 0],
+    'chiral_tag': [0, 1, 2, 3],
+    'num_Hs': [0, 1, 2, 3, 4],
+    'hybridization': [
+        OEHybridization_sp,
+        OEHybridization_sp2,
+        OEHybridization_sp3,
+        OEHybridization_sp3d,
+        OEHybridization_sp3d2
+    ],
+}
+ATOM_FDIM = sum(len(choices) + 1 for choices in ATOM_FEATURES.values()) + 2
+
+# Timing sensitve import with Globals
+from .featurization_utils import *
+
+def get_atom_fdim(args: Namespace) -> int:
+    """
+    Gets the dimensionality of atom features.
+
+    :param: Arguments.
+    """
+    return ATOM_FDIM
+
+def atom_features(atom: OEAtomBase, functional_groups: List[int] = None) -> List[Union[bool, int, float]]:
+    """
+    Builds a feature vector for an atom.
+
+    :param atom: An RDKit atom.
+    :param functional_groups: A k-hot vector indicating the functional groups the atom belongs to.
+    :return: A list containing the atom features.
+    """
+    features = onek_encoding_unk(atom.GetAtomicNum() - 1, ATOM_FEATURES['atomic_num']) + \
+               onek_encoding_unk(atom.GetDegree(), ATOM_FEATURES['degree']) + \
+               onek_encoding_unk(atom.GetFormalCharge(), ATOM_FEATURES['formal_charge']) + \
+               onek_encoding_unk(int(atom.IsChiral()), ATOM_FEATURES['chiral_tag']) + \
+               onek_encoding_unk(int(atom.GetTotalHCount()), ATOM_FEATURES['num_Hs']) + \
+               onek_encoding_unk(int(atom.GetHyb()), ATOM_FEATURES['hybridization']) + \
+               [1 if atom.IsAromatic() else 0] + \
+               [round(OEGetAverageWeight(atom.GetAtomicNum()) * 0.01,
+                      6)]  # scaled to about the same range as other features
+
+    if functional_groups is not None:
+        features += functional_groups
+    return features
+
+
+def bond_features(bond: OEBondBase) -> List[Union[bool, int, float]]:
+    """
+    Builds a feature vector for a bond.
+
+    :param bond: A RDKit bond.
+    :return: A list containing the bond features.
+    """
+    if bond is None:
+        fbond = [1] + [0] * (BOND_FDIM - 1)
+    else:
+        bt = bond.GetOrder()
+        fbond = [
+            0,  # bond is not None
+            bt == 1,
+            bt == 2,
+            bt == 3,
+            bt == bond.IsAromatic(),
+            (0),  # I don't see conjugation as an OE property.  May be more work to extract
+            (bond.IsInRing() if bt is not None else 0)
+        ]
+        nb1 = bond.GetBgn()
+        nb2 = bond.GetEnd()
+        st = bond.GetStereo([nb1, nb2], OEAtomStereo_Tetrahedral)
+        fbond += onek_encoding_unk(int(st), list(range(6)))
+    return fbond
+
+
+class MolGraph:
+    """
+    A MolGraph represents the graph structure and featurization of a single molecule.
+
+    A MolGraph computes the following attributes:
+    - smiles: Smiles string.
+    - n_atoms: The number of atoms in the molecule.
+    - n_bonds: The number of bonds in the molecule.
+    - f_atoms: A mapping from an atom index to a list atom features.
+    - f_bonds: A mapping from a bond index to a list of bond features.
+    - a2b: A mapping from an atom index to a list of incoming bond indices.
+    - b2a: A mapping from a bond index to the index of the atom the bond originates from.
+    - b2revb: A mapping from a bond index to the index of the reverse bond.
+    """
+
+    def __init__(self, smiles: str, args: Namespace):
+        """
+        Computes the graph structure and featurization of a molecule.
+
+        :param smiles: A smiles string.
+        :param args: Arguments.
+        """
+        self.smiles = smiles
+        self.n_atoms = 0  # number of atoms
+        self.n_bonds = 0  # number of bonds
+        self.f_atoms = []  # mapping from atom index to atom features
+        self.f_bonds = []  # mapping from bond index to concat(in_atom, bond) features
+        self.a2b = []  # mapping from atom index to incoming bond indices
+        self.b2a = []  # mapping from bond index to the index of the atom the bond is coming from
+        self.b2revb = []  # mapping from bond index to the index of the reverse bond
+
+        # Convert smiles to molecule
+        mol = OEGraphMol()
+        OESmilesToMol(mol, self.smiles)
+        OEAssignHybridization(mol)
+        OEPerceiveChiral(mol)
+
+        # mol = Chem.MolFromSmiles(smiles)
+
+        # fake the number of "atoms" if we are collapsing substructures
+        self.n_atoms = mol.NumAtoms()
+        self.atom_idx = []
+
+        # Get atom features
+        for atom in mol.GetAtoms():
+            self.f_atoms.append(atom_features(atom))
+            self.atom_idx.append(atom)
+            # self.f_atoms is a list of lists features for each atom
+
+        self.f_atoms = [self.f_atoms[i] for i in range(self.n_atoms)]
+        # I don't think this does a thing?
+
+        for _ in range(self.n_atoms):
+            self.a2b.append([])
+
+        # Get bond features
+        for a1 in range(len(self.atom_idx)):
+            for a2 in range(a1 + 1, len(self.atom_idx)):
+                bond = mol.GetBond(self.atom_idx[a1], self.atom_idx[a2])
+
+                if bond is None:
+                    continue
+
+                f_bond = bond_features(bond)
+
+                if args.atom_messages:
+                    self.f_bonds.append(f_bond)
+                    self.f_bonds.append(f_bond)
+                else:
+                    self.f_bonds.append(self.f_atoms[a1] + f_bond)
+                    self.f_bonds.append(self.f_atoms[a2] + f_bond)
+
+                # Update index mappings
+                b1 = self.n_bonds
+                b2 = b1 + 1
+                self.a2b[a2].append(b1)  # b1 = a1 --> a2
+                self.b2a.append(a1)
+                self.a2b[a1].append(b2)  # b2 = a2 --> a1
+                self.b2a.append(a2)
+                self.b2revb.append(b2)
+                self.b2revb.append(b1)
+                self.n_bonds += 2
