+++ chemprop/features/__init__.py
@@ -1,3 +1,15 @@
+try:
+    import openeye
+    from .OEMolGraph import atom_features, bond_features, MolGraph, get_atom_fdim
+except ImportError:
+    from .RDKitMolGraph import atom_features, bond_features, MolGraph, get_atom_fdim
+
+from .featurization import BatchMolGraph, mol2graph
+from .featurization_utils import get_bond_fdim, clear_cache, SMILES_TO_GRAPH
 from .features_generators import get_available_features_generators, get_features_generator
-from .featurization import atom_features, bond_features, BatchMolGraph, get_atom_fdim, get_bond_fdim, mol2graph, clear_cache
 from .utils import load_features, save_features
+
+
+
+
+
