+++ chemprop/features/featurization_utils.py
@@ -1,0 +1,46 @@
+# Stuff for typing:
+from argparse import Namespace
+from typing import List, Tuple, Union
+
+
+# Distance feature sizes
+PATH_DISTANCE_BINS = list(range(10))
+THREE_D_DISTANCE_MAX = 20
+THREE_D_DISTANCE_STEP = 1
+THREE_D_DISTANCE_BINS = list(range(0, THREE_D_DISTANCE_MAX + 1, THREE_D_DISTANCE_STEP))
+
+BOND_FDIM = 14
+
+# Memoization
+SMILES_TO_GRAPH = {}
+
+
+def clear_cache():
+    """Clears featurization cache."""
+    global SMILES_TO_GRAPH
+    SMILES_TO_GRAPH = {}
+
+
+def get_bond_fdim(args: Namespace) -> int:
+    """
+    Gets the dimensionality of bond features.
+
+    :param: Arguments.
+    """
+    return BOND_FDIM
+
+
+def onek_encoding_unk(value: int, choices: List[int]) -> List[int]:
+    """
+    Creates a one-hot encoding.
+
+    :param value: The value for which the encoding should be one.
+    :param choices: A list of possible values.
+    :return: A one-hot encoding of the value in a list of length len(choices) + 1.
+    If value is not in the list of choices, then the final element in the encoding is 1.
+    """
+    encoding = [0] * (len(choices) + 1)
+    index = choices.index(value) if value in choices else -1
+    encoding[index] = 1
+
+    return encoding
